@@ -7,33 +7,33 @@ namespace DigipolisGent\API\Client\Request;
 /**
  * The different Method types.
  */
-class MethodType
+final class MethodType
 {
     /**
      * GET sending method.
      *
      * @var string
      */
-    const GET = 'GET';
+    public const GET = 'GET';
 
     /**
      * POST sending method.
      *
      * @var string
      */
-    const POST = 'POST';
+    public const POST = 'POST';
 
     /**
      * PUT sending method.
      *
      * @var string
      */
-    const PUT = 'PUT';
+    public const PUT = 'PUT';
 
     /**
      * DELETE sending method.
      *
      * @var string
      */
-    const DELETE  = 'DELETE';
+    public const DELETE  = 'DELETE';
 }
